requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
et_xmlfile==2.0.0
fpdf==1.7.2
idna==3.10
lxml==5.3.0
openpyxl==3.1.5
pillow==11.0.0
python-docx==1.1.2
python-pptx==1.0.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
XlsxWriter==3.2.0
